pyproject.toml
[tool.poetry]
name = "gene-fetch" 
version = "1.0.19"
description = "Gene Fetch: High-throughput NCBI Sequence Retrieval Tool"
authors = ["D. Parsons <d.parsons@nhm.ac.uk>, B. Price <b.price@nhm.ac.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gene_fetch", from = "src"}]
include = ["tests/*"]
repository = "https://github.com/bge-barcoding/gene_fetch"
keywords = ["bioinformatics", "ncbi", "sequence", "genomics", "taxonomy", "barcodes"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[tool.poetry.dependencies]
python = ">=3.9"
biopython = ">=1.80"
ratelimit = ">=2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
flake8 = "^6.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.poetry.scripts]
gene-fetch = "gene_fetch.main:main"

[tool.poetry.urls]
"Homepage" = "https://github.com/bge-barcoding/gene_fetch"
"Bug Tracker" = "https://github.com/bge-barcoding/gene_fetch/issues"
